+++ api_yamdb/reviews/constants.py
@@ -2,3 +2,4 @@
 
 NAME_MAX_LENGTH = 256
 SLUG_MAX_LENGTH = 50
+OBJECT_NAME_DISPLAY_LENGTH = 20
